refactor: create top-level `commands` and `events` mods
