+++ packages/wm/src/commands/general/disable_binding_mode.rs
