+++ packages/wm/src/commands/general/enable_binding_mode.rs
