+++ packages/wm/src/commands/general/shell_exec.rs
