+++ packages/wm/src/commands/general/toggle_pause.rs
