+++ packages/wm/src/events/handle_display_settings_changed.rs
