+++ packages/wm/src/events/handle_window_destroyed.rs
