+++ packages/wm/src/events/handle_window_hidden.rs
