+++ packages/wm/src/events/handle_window_location_changed.rs
