+++ packages/wm/src/events/handle_window_minimize_ended.rs
