+++ packages/wm/src/events/handle_window_minimized.rs
