+++ packages/wm/src/events/handle_window_shown.rs
