+++ packages/wm/src/events/handle_window_title_changed.rs
